requirements.txt
langchain==0.3.7
langchain_community==0.3.5
langchain_core==0.3.15
langchain_openai==0.2.6
python-dotenv==1.0.1
streamlit==1.40.0
